# Inline getconstant on JIT