--- a/mjit.c
+++ b/mjit.c
@@ -82,6 +82,24 @@ mjit_gc_exit_hook(void)
     CRITICAL_SECTION_FINISH(4, "mjit_gc_exit_hook");
 }
 
+// Lock setinlinecache
+void
+rb_mjit_before_vm_ic_update(void)
+{
+    if (!mjit_enabled)
+        return;
+    CRITICAL_SECTION_START(3, "before vm_ic_update");
+}
+
+// Unlock setinlinecache
+void
+rb_mjit_after_vm_ic_update(void)
+{
+    if (!mjit_enabled)
+        return;
+    CRITICAL_SECTION_FINISH(3, "after vm_ic_update");
+}
+
 // Deal with ISeq movement from compactor
 void
 mjit_update_references(const rb_iseq_t *iseq)
@@ -374,6 +392,14 @@ void
 rb_mjit_recompile_inlining(const rb_iseq_t *iseq)
 {
     rb_mjit_iseq_compile_info(iseq->body)->disable_inlining = true;
+    mjit_recompile(iseq);
+}
+
+// Recompile iseq, disabling getconstant inlining
+void
+rb_mjit_recompile_const(const rb_iseq_t *iseq)
+{
+    rb_mjit_iseq_compile_info(iseq->body)->disable_const_cache = true;
     mjit_recompile(iseq);
 }