## VM
Just to confirm there's no significant impact:

```
$ benchmark-driver -v --rbenv 'before;after' --repeat-count=4 --alternate --output=all benchmark.yml
before: ruby 3.0.0dev (2020-12-15T07:35:45Z master 5463eff5f6) [x86_64-linux]
after: ruby 3.0.0dev (2020-12-15T08:13:18Z getconstant 0307b7549f) [x86_64-linux]
Calculating -------------------------------------
                                   before                 after
Optcarrot 180 frames    44.43906567133481     44.08993804595956 fps
                        44.85131944661266     44.40423525746272
                        45.29782481779199     44.85121796957385
                        45.36123098826729     45.67076168086072
```

## JIT
```
$ benchmark-driver -v --rbenv 'before --jit;after --jit' --repeat-count=12 --alternate --output=all benchmark.yml
before --jit: ruby 3.0.0dev (2020-12-15T07:35:45Z master 5463eff5f6) +JIT [x86_64-linux]
after --jit: ruby 3.0.0dev (2020-12-15T08:13:18Z getconstant 0307b7549f) +JIT [x86_64-linux]
Calculating -------------------------------------
                             before --jit           after --jit
Optcarrot 180 frames    63.53308278121263     81.96004198024043 fps
                        79.09523835603413     82.55712437192619
                        81.11530369952305     86.71056868058092
                        83.44053974954316     87.62311777724165
                        83.78691925272057     87.67412812654518
                        84.15543609932206     88.75210458341785
                        84.39967476701766     88.87561592203616
                        86.64669158401497     89.87739375355882
                        86.79082003391167     89.97637400528664
                        86.94517770345354     89.97822437196872
                        87.47667920072629     90.05956719853972
                        87.57852046989133     90.24818250184856
```

```
$ benchmark-driver -v --rbenv 'before --jit;after --jit' --repeat-count=2 --alternate --output=all benchmark_3000.yml
before --jit: ruby 3.0.0dev (2020-12-15T07:35:45Z master 5463eff5f6) +JIT [x86_64-linux]
after --jit: ruby 3.0.0dev (2020-12-15T08:13:18Z getconstant 0307b7549f) +JIT [x86_64-linux]
Calculating -------------------------------------
                              before --jit           after --jit
Optcarrot 3000 frames     94.5293680511305      99.3363361451684 fps
                          98.5481222681076      99.3946139186733
```